+++ diplomacy/src/judge/retreat/resolver.rs
@@ -144,16 +144,3 @@ impl<O> OrderOutcome<O> {
         }
     }
 }
-
-/// Most `DestStatus` values block a retreat-phase move order from succeeding or exerting
-/// influence on the move destination. These values can appear in the `InvalidDestination`
-/// variant of `OrderOutcome`. Note that `DestStatus::Available` will never equal an order outcome.
-impl<O> PartialEq<DestStatus> for OrderOutcome<O> {
-    fn eq(&self, other: &DestStatus) -> bool {
-        if let OrderOutcome::InvalidDestination(status) = self {
-            *other != DestStatus::Available && status == other
-        } else {
-            false
-        }
-    }
-}